Order Events by time/data fields.
Events can be ordered by the time or any of its data fields by adding an
'order=field' parameter to the query. 'order=-time' is the default.

Descending order can be specified by prepending "-" to the field name.

If the ordering field is "time" the field name is transformed to "t",
otherwise it is transformed to "d.{field_name}".